propolis-standalone-config needn't be a crate
Since config extraction is not part of `savex` in propolis-utils
anymore, the configuration definitions for propolis-standalone do not
need to be exposed outside of the binary crate itself.
