+++ bin/propolis-standalone/Cargo.toml
@@ -25,7 +25,7 @@ tokio = { workspace = true, features = ["io-util", "rt-multi-thread"] }
 serde = { workspace = true, features = ["derive"] }
 propolis.workspace = true
 crucible-client-types = { workspace = true, optional = true }
-propolis-standalone-config = { workspace = true }
+cpuid_profile_config.workspace = true
 erased-serde.workspace = true
 serde_json.workspace = true
 slog.workspace = true
