+++ bin/propolis-utils/Cargo.toml
@@ -22,4 +22,3 @@ propolis = { workspace = true, default-features = false }
 bhyve_api = { workspace = true }
 libc = { workspace = true }
 serde_json.workspace = true
-propolis-standalone-config = { workspace = true }
