--- crates/propolis-standalone-config/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "propolis-standalone-config"
-version = "0.0.0"
-license = "MPL-2.0"
-edition = "2021"
-
-[lib]
-test = false
-doctest = false
-
-[dependencies]
-cpuid_profile_config.workspace = true
-serde.workspace = true
-serde_derive.workspace = true
-toml.workspace = true
